feat: Add const constructor to PlaceLocation class
The code changes add a const constructor to the `PlaceLocation` class in the `place.dart` file. This allows for the creation of `PlaceLocation` objects with constant values for latitude, longitude, and address.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.