feat: Add onTap functionality to select location on the map
The code changes introduce the onTap functionality to the GoogleMap widget in the _MapScreenState class. When the user taps on the map, the _pickedLocation variable is updated with the tapped position. This allows for the selection of a location on the map.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.