feat: Add sqflite plugin to GeneratedPluginRegistrant.swift and pubspec.yaml
The code changes add the sqflite plugin to the GeneratedPluginRegistrant.swift file in the macos/Flutter directory and the pubspec.yaml file. This allows for the integration of sqflite functionality in the app.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.