feat: Add lazy loading for images in UserPlacesNotifier
The code changes in the `UserPlacesNotifier` class introduce lazy loading for images. This is achieved by using the `path_provider` and `path` packages to get the application documents directory and copy the image file to that directory. The copied image is then assigned to the `newPlace` object before adding it to the list of places.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.