feat: Refactor Place constructor to allow for optional id parameter
The code changes in the `Place` class refactor the constructor to allow for an optional `id` parameter. This provides flexibility in assigning an existing id or generating a new one using the `uuid` package. This change ensures that the `id` property is not required when creating a new `Place` object.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.