feat: Update PlacesScreen to use lazy loading for images
The code changes in the `PlacesScreen` class introduce lazy loading for images. This is achieved by using the `FutureBuilder` widget to asynchronously load the places data and display a loading indicator while the data is being fetched. Once the data is available, the `PlacesList` widget is rendered with the loaded places.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.