feat: Add .cache_ggshield file
The code changes add a new file, `.cache_ggshield`, which stores information about the last found secrets. This file is used by the `ggshield` tool for secret scanning.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.